Reorganize packages for TrackScheme styles.
