+++ src/main/java/org/mastodon/revised/trackscheme/display/TrackSchemePanel.java
@@ -29,8 +29,7 @@ import org.mastodon.revised.trackscheme.TrackSchemeSelection;
 import org.mastodon.revised.trackscheme.TrackSchemeVertex;
 import org.mastodon.revised.trackscheme.display.TrackSchemeOptions.Values;
 import org.mastodon.revised.trackscheme.display.animate.AbstractAnimator;
-import org.mastodon.revised.trackscheme.display.laf.DefaultTrackSchemeOverlay;
-import org.mastodon.revised.trackscheme.display.laf.TrackSchemeStyle;
+import org.mastodon.revised.trackscheme.display.style.TrackSchemeStyle;
 import org.mastodon.revised.ui.selection.FocusListener;
 import org.mastodon.revised.ui.selection.HighlightListener;
 import org.mastodon.revised.ui.selection.NavigationEtiquette;
