+++ src/main/java/org/mastodon/revised/trackscheme/display/style/TrackSchemeStyleChooserDialog.java
@@ -1,4 +1,4 @@
-package org.mastodon.revised.trackscheme.display.ui;
+package org.mastodon.revised.trackscheme.display.style;
 
 import java.awt.BorderLayout;
 import java.awt.Container;
@@ -35,11 +35,10 @@ import org.mastodon.revised.trackscheme.TrackSchemeNavigation;
 import org.mastodon.revised.trackscheme.TrackSchemeSelection;
 import org.mastodon.revised.trackscheme.display.TrackSchemeOptions;
 import org.mastodon.revised.trackscheme.display.TrackSchemePanel;
-import org.mastodon.revised.trackscheme.display.laf.TrackSchemeStyle;
-import org.mastodon.revised.trackscheme.display.ui.dummygraph.DummyEdge;
-import org.mastodon.revised.trackscheme.display.ui.dummygraph.DummyGraph;
-import org.mastodon.revised.trackscheme.display.ui.dummygraph.DummyGraph.Examples;
-import org.mastodon.revised.trackscheme.display.ui.dummygraph.DummyVertex;
+import org.mastodon.revised.trackscheme.display.style.dummygraph.DummyEdge;
+import org.mastodon.revised.trackscheme.display.style.dummygraph.DummyGraph;
+import org.mastodon.revised.trackscheme.display.style.dummygraph.DummyVertex;
+import org.mastodon.revised.trackscheme.display.style.dummygraph.DummyGraph.Examples;
 import org.mastodon.revised.ui.grouping.GroupManager;
 import org.mastodon.revised.ui.selection.FocusModel;
 import org.mastodon.revised.ui.selection.HighlightModel;
